requirements.txt
Django==5.0.1
mysqlclient==2.2.1
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2.0.0
python-decouple==3.8
